add final changes to new  repository

+++ src/service/clienteService.js
@@ -1,11 +1,13 @@
-const clienteRepository = require("../repository/clienteRepository").default;
+const clienteRepository = require("../repository/clienteRepository");
 const ClienteDTO = require("../dto/clienteDTO");
 
 class ClienteService {
   async createCliente(clienteDTO) {
-    const { nome, email, pedido } = clienteDTO;
+    const { nome, email, pedidoId } = clienteDTO;
 
-    return await clienteRepository.create(new ClienteDTO(nome, email, pedido));
+    return await clienteRepository.create(
+      new ClienteDTO(nome, email, pedidoId)
+    );
   }
 
   async getAllClientes() {
@@ -21,6 +23,11 @@ class ClienteService {
   }
 
   async deleteCliente(clienteId) {
+    const cliente = await clienteRepository.getById(clienteId);
+
+    if (!cliente) {
+      throw new Error("Cliente não encontrado.");
+    }
     return clienteRepository.delete(clienteId);
   }
 }
